etl/app/requirements.txt
django==3.2
uwsgi==2.0.20
flake8==4.0
python-dotenv==0.21.0
django-split-settings==1.2.0
psycopg2==2.9.5
django-debug-toolbar==3.4
djangorestframework==3.14.0